GroundStationFlask/requirements.txt
flask
flask-cors
flask-queue-sse
gunicorn
gevent
tinydb
tinyrecord
pyserial